#37 fix in casting of transformer outputs

+++ skorecard/preprocessing/binning1d.py
@@ -96,7 +96,7 @@ class BucketTransformer(BaseEstimator, TransformerMixin):
         for ix in self.boundary_dict.keys():
             X[:, ix] = np.digitize(X[:, ix], self.boundary_dict[ix][1:], right=True)
 
-        return X
+        return X.astype(int)
 
 
 class SimpleBucketTransformer(BucketTransformer):
